+++ synapse-core/src/main/java/de/otto/synapse/state/ConcurrentHashMapStateRepository.java
@@ -2,7 +2,7 @@ package de.otto.synapse.state;
 
 import java.util.concurrent.ConcurrentHashMap;
 
-public class ConcurrentHashMapStateRepository<V> extends StateRepository<V> {
+public class ConcurrentHashMapStateRepository<V> extends ConcurrentMapStateRepository<V> {
 
     public ConcurrentHashMapStateRepository() {
         super(new ConcurrentHashMap<>());